feat: 그룹 맴버 - 그룹 탈퇴 기능 구현 그룹 탈퇴시 그룹장은 탈퇴를 못하는 로직 생성 (#147)

+++ src/main/java/net/diveon/backend/domain/group/service/GroupMemberService.java
@@ -14,6 +14,7 @@ import net.diveon.backend.domain.group.repository.GroupUserRepository;
 import net.diveon.backend.domain.user.entity.User;
 import net.diveon.backend.domain.user.repository.UserRepository;
 import net.diveon.backend.global.exception.GroupAssignRequestNotPendingException;
+import net.diveon.backend.global.exception.GroupLeaderCannotLeaveException;
 import net.diveon.backend.global.exception.GroupNotFoundException;
 import net.diveon.backend.global.exception.GroupUserNotFoundException;
 import net.diveon.backend.global.exception.UserNotFoundException;
@@ -105,6 +106,10 @@ public class GroupMemberService {
 
         GroupUser groupUser = groupUserRepository.findByGroupIdAndUserId(groupId, userId)
                 .orElseThrow(GroupUserNotFoundException::new);
+        //양도 만약에 생각하면 이게 더 좋은 검사인듯? 바꿀수는 있습니다 userId랑 그룹장 leaderId 검사하게
+        if (groupUser.getRole() == GroupRole.LEADER) {
+            throw new GroupLeaderCannotLeaveException();
+        }
 
         groupUserRepository.delete(groupUser);
         return new GroupMemberCommonResponse(userId, "none");
